+++ Editor/d4rkAvatarOptimizer.cs
@@ -138,6 +138,7 @@ public class d4rkAvatarOptimizer : MonoBehaviour
             keepTransforms.Clear();
             convertedMeshRendererPaths.Clear();
             constantAnimatedValuesToAdd.Clear();
+            animatedMaterialPropertyDefaultValues.Clear();
             ClearCaches();
             DisplayProgressBar("Destroying unused components", 0.2f);
             Profiler.StartNextSection("DestroyEditorOnlyGameObjects()");
@@ -430,6 +431,7 @@ public class d4rkAvatarOptimizer : MonoBehaviour
     private Dictionary<(string path, int index), (string path, int index)> materialSlotRemap = new Dictionary<(string, int), (string, int)>();
     private Dictionary<string, HashSet<string>> animatedMaterialProperties = new Dictionary<string, HashSet<string>>();
     private Dictionary<string, HashSet<string>> fusedAnimatedMaterialProperties = new Dictionary<string, HashSet<string>>();
+    private Dictionary<string, Dictionary<string, Vector4>> animatedMaterialPropertyDefaultValues = new Dictionary<string, Dictionary<string, Vector4>>();
     private List<List<Texture2D>> textureArrayLists = new List<List<Texture2D>>();
     private List<Texture2DArray> textureArrays = new List<Texture2DArray>();
     private Dictionary<Material, List<(string name, Texture2DArray array)>> texArrayPropertiesToSet = new Dictionary<Material, List<(string name, Texture2DArray array)>>();
@@ -2090,6 +2092,36 @@ public class d4rkAvatarOptimizer : MonoBehaviour
         return cache_GetAllUsedFXLayerAnimationClips = usedClips;
     }
 
+    private bool? cache_DoesFXLayerUseWriteDefaults = null;
+    public bool DoesFXLayerUseWriteDefaults()
+    {
+        if (cache_DoesFXLayerUseWriteDefaults != null)
+            return cache_DoesFXLayerUseWriteDefaults.Value;
+        var fxLayer = GetFXLayer();
+        if (fxLayer == null)
+            return false;
+        var fxLayerLayers = GetFXLayerLayers();
+        for (int i = 0; i < fxLayerLayers.Length; i++)
+        {
+            var stateMachine = fxLayerLayers[i].stateMachine;
+            if (stateMachine == null || (OptimizeFXLayer && IsMergeableFXLayer(i)))
+                continue;
+            foreach (var state in stateMachine.EnumerateAllStates())
+            {
+                if (state.motion is BlendTree blendTree && blendTree.blendType == BlendTreeType.Direct)
+                    continue;
+                if (state.writeDefaultValues)
+                {
+                    cache_DoesFXLayerUseWriteDefaults = true;
+                    return true;
+                }
+            }
+        }
+        cache_DoesFXLayerUseWriteDefaults = false;
+        return false;
+    }
+
+
     private HashSet<AnimationClip> cache_GetAllUsedAnimationClips = null;
     private HashSet<AnimationClip> GetAllUsedAnimationClips()
     {
@@ -2329,10 +2361,9 @@ public class d4rkAvatarOptimizer : MonoBehaviour
                 if (!optimizedMaterials.TryGetValue(material, out var optimizedMaterial))
                 {
                     DisplayProgressBar("Optimizing swap material: " + material.name);
-                    var matWrapper = new List<List<Material>>() { new List<Material>() { material } };
-                    var sourcePathWrapper = new List<List<string>>() { Enumerable.Repeat(entry.Key.path, mergedMeshCount).ToList() };
+                    var matWrapper = new List<List<(Material, List<string>)>>() { new List<(Material, List<string>)>() { (material, new List<string> { entry.Key.path } ) } };
                     var mergedMeshIndexWrapper = new List<List<int>>() { new List<int>() { meshIndex } };
-                    optimizedMaterials[material] = CreateOptimizedMaterials(matWrapper, mergedMeshCount, targetPath, sourcePathWrapper, mergedMeshIndexWrapper)[0];
+                    optimizedMaterials[material] = CreateOptimizedMaterials(matWrapper, mergedMeshCount, targetPath, mergedMeshIndexWrapper)[0];
                 }
             }
         }
@@ -3167,27 +3198,48 @@ public class d4rkAvatarOptimizer : MonoBehaviour
         return name + " " + count;
     }
 
+    private Dictionary<string, Material> cache_GetFirstMaterialOnPath = null;
+    public Material GetFirstMaterialOnPath(string path)
+    {
+        if (cache_GetFirstMaterialOnPath == null)
+            cache_GetFirstMaterialOnPath = new Dictionary<string, Material>();
+        if (cache_GetFirstMaterialOnPath.TryGetValue(path, out var mat))
+            return mat;
+        var renderer = GetTransformFromPath(path)?.GetComponent<Renderer>();
+        if (renderer == null || renderer.sharedMaterials.Length == 0)
+            return cache_GetFirstMaterialOnPath[path] = null;
+        return cache_GetFirstMaterialOnPath[path] = renderer.sharedMaterials[0];
+    }
+
     private Material[] CreateOptimizedMaterials(
-        List<List<Material>> sources,
+        List<List<(Material mat, List<string> paths)>> sources,
         int meshToggleCount,
         string path,
-        List<List<string>> originalMeshPaths = null,
         List<List<int>> mergedMeshIndices = null)
     {
         if (!(WritePropertiesAsStaticValues || sources.Any(list => list.Count > 1) || (meshToggleCount > 1 && MergeSkinnedMeshesWithShaderToggle)))
         {
-            return sources.Select(list => list[0]).ToArray();
+            return sources.Select(list => list[0].mat).ToArray();
         }
         if (!fusedAnimatedMaterialProperties.TryGetValue(path, out var usedMaterialProps))
             usedMaterialProps = new HashSet<string>();
         if (mergedMeshIndices == null)
             mergedMeshIndices = sources.Select(s => Enumerable.Range(0, meshToggleCount).ToList()).ToList();
         HashSet<(string name, bool isVector)> defaultAnimatedProperties = null;
+        var animatedPropertyOnMeshID = new Dictionary<string, bool[]>();
         oldPathToMergedPaths.TryGetValue(path, out var allOriginalMeshPaths);
         var sameAnimatedProperties = GetSameAnimatedPropertiesOnMergedMesh(path);
+        var originalMeshPaths = sources.Select(l => l.SelectMany(t => t.paths).Distinct().ToList()).ToList();
         if (allOriginalMeshPaths != null && (sources.Count != 1 || sources[0].Count != 1)) {
             defaultAnimatedProperties = new HashSet<(string name, bool isVector)>();
             for (int i = 0; i < allOriginalMeshPaths.Count; i++) {
+                Dictionary<string, Vector4> defaultValuesForCurrentPath = null;
+                Material defaultMaterialForCurrentPath = GetFirstMaterialOnPath(allOriginalMeshPaths[i][0]);
+                if (!animatedMaterialPropertyDefaultValues.TryGetValue(path, out defaultValuesForCurrentPath))
+                {
+                    defaultValuesForCurrentPath = new Dictionary<string, Vector4>();
+                    animatedMaterialPropertyDefaultValues[path] = defaultValuesForCurrentPath;
+                }
                 if (animatedMaterialProperties.TryGetValue(allOriginalMeshPaths[i][0], out var animatedProps)) {
                     foreach (var prop in animatedProps) {
                         string name = prop;
@@ -3203,11 +3255,24 @@ public class d4rkAvatarOptimizer : MonoBehaviour
                         }
                         defaultAnimatedProperties.Add(($"d4rkAvatarOptimizer{name}_ArrayIndex{i}", isVector));
                         defaultAnimatedProperties.Add((name, isVector));
+                        if (!animatedPropertyOnMeshID.TryGetValue(name, out var animatedOnMesh)) {
+                            animatedOnMesh = new bool[allOriginalMeshPaths.Count];
+                            animatedPropertyOnMeshID[name] = animatedOnMesh;
+                        }
+                        animatedOnMesh[i] = true;
+                        if (defaultMaterialForCurrentPath != null && defaultMaterialForCurrentPath.HasProperty(name)) 
+                        {
+                            defaultValuesForCurrentPath[$"d4rkAvatarOptimizer{name}_ArrayIndex{i}"] = isVector
+                                ? defaultMaterialForCurrentPath.GetVector(name)
+                                : new Vector4(defaultMaterialForCurrentPath.GetFloat(name), 0, 0, 0);
+                        }
                     }
                 }
                 defaultAnimatedProperties.Add(($"_IsActiveMesh{i}", false));
             }
         }
+        if (!DoesFXLayerUseWriteDefaults())
+            animatedPropertyOnMeshID = null;
         var materials = new Material[sources.Count];
         var parsedShader = new ParsedShader[sources.Count];
         var sanitizedMaterialNames = new string[sources.Count];
@@ -3222,7 +3287,7 @@ public class d4rkAvatarOptimizer : MonoBehaviour
         var stripShadowVariants = new bool[sources.Count];
         for (int i = 0; i < sources.Count; i++)
         {
-            var source = sources[i];
+            var source = sources[i].Select(t => t.mat).ToList();
             parsedShader[i] = ShaderAnalyzer.Parse(source[0]?.shader);
             if (parsedShader[i] == null || !parsedShader[i].parsedCorrectly)
             {
@@ -3413,14 +3478,16 @@ public class d4rkAvatarOptimizer : MonoBehaviour
                     setShaderKeywords[i],
                     poiUsedPropertyDefines[i],
                     sanitizedMaterialNames[i],
-                    stripShadowVariants[i]);
+                    stripShadowVariants[i],
+                    animatedPropertyOnMeshID
+                );
             }
         });
         Profiler.EndSection();
 
         for (int i = 0; i < sources.Count; i++)
         {
-            var source = sources[i];
+            var source = sources[i].Select(t => t.mat).ToList();
             if (parsedShader[i] == null || !parsedShader[i].parsedCorrectly)
                 continue;
 
@@ -3566,29 +3633,38 @@ public class d4rkAvatarOptimizer : MonoBehaviour
                     foreach (var animPropName in animatedProperties)
                     {
                         var propName = animPropName;
-                        bool isVector = propName.EndsWith(".x");
-                        bool isColor = propName.EndsWith(".r");
-                        if (isColor || isVector) {
+                        bool isVector = propName.EndsWith(".x") || propName.EndsWith(".r");
+                        if (isVector) {
                             propName = propName.Substring(0, propName.Length - 2);
                         } else if (propName[propName.Length - 2] == '.') {
                             continue;
-                        } else if (animatedProperties.Contains($"{propName}.x")) {
+                        } else if (animatedProperties.Contains($"{propName}.x") || animatedProperties.Contains($"{propName}.r")) {
                             isVector = true;
-                        } else if (animatedProperties.Contains($"{propName}.r")) {
-                            isColor = true;
                         }
                         for (int mID = 0; mID < meshCount; mID++)
                         {
                             var propArrayName = $"d4rkAvatarOptimizer{propName}_ArrayIndex{mID}";
                             if (!mat.HasProperty(propArrayName))
                                 continue;
-                            var signal = float.NaN;
+                            var signal = DoesFXLayerUseWriteDefaults() ? 0.0f : float.NaN;
                             if (isVector) {
                                 mat.SetVector(propArrayName, new Vector4(signal, signal, signal, signal));
-                            } else if(isColor) {
-                                mat.SetColor(propArrayName, new Color(signal, signal, signal, signal));
                             } else {
                                 mat.SetFloat(propArrayName, signal);
+                            }
+                        }
+                    }
+                    if (DoesFXLayerUseWriteDefaults() && animatedMaterialPropertyDefaultValues.TryGetValue(GetPathToRoot(meshRenderer), out var defaultValues))
+                    {
+                        foreach (var defaultProp in defaultValues)
+                        {
+                            if (mat.HasFloat(defaultProp.Key))
+                            {
+                                mat.SetFloat(defaultProp.Key, defaultProp.Value.x);
+                            }
+                            else if (mat.HasVector(defaultProp.Key))
+                            {
+                                mat.SetVector(defaultProp.Key, defaultProp.Value);
                             }
                         }
                     }
@@ -3758,7 +3834,7 @@ public class d4rkAvatarOptimizer : MonoBehaviour
                 }
             }
             var toOptimize = mats.Select(t => t.material).Where(m => !alreadyOptimizedMaterials.Contains(m)).Distinct().ToList();
-            var optimizeMaterialWrapper = toOptimize.Select(m => new List<Material>() { m }).ToList();
+            var optimizeMaterialWrapper = toOptimize.Select(m => new List<(Material, List<string>)>() { (m, new List<string> { path } ) }).ToList();
             var optimizedMaterialsList = CreateOptimizedMaterials(optimizeMaterialWrapper, 0, GetPathToRoot(meshRenderer));
             var optimizedMaterials = toOptimize.Select((mat, index) => (mat, index))
                 .ToDictionary(t => t.mat, t => optimizedMaterialsList[t.index]);
@@ -4042,13 +4118,15 @@ public class d4rkAvatarOptimizer : MonoBehaviour
                 return remap;
             }
 
-            var originalMeshPaths = matchedSlots.Select(list => list.Select(slot => GetOriginalSlot((meshPath, slot.index)).path).Distinct().ToList()).ToList();
-            var uniqueMatchedMaterials = uniqueMatchedSlots.Select(list => list.Select(slot => slot.material).ToList()).ToList();
-            var optimizedMaterials = CreateOptimizedMaterials(uniqueMatchedMaterials, meshCount > 1 ? meshCount : 0, meshPath, originalMeshPaths, mergedMeshIndices);
+            var allSlots = matchedSlots.SelectMany(list => list).ToList();
+            var uniqueMatchedMaterials = uniqueMatchedSlots.Select(list => list.Select(slot =>
+                (slot.material, allSlots.Where(slot2 => slot2.material == slot.material).Select(slot2 => GetOriginalSlot((meshPath, slot2.index)).path).ToList())
+            ).ToList()).ToList();
+            var optimizedMaterials = CreateOptimizedMaterials(uniqueMatchedMaterials, meshCount > 1 ? meshCount : 0, meshPath, mergedMeshIndices);
 
             for (int i = 0; i < uniqueMatchedMaterials.Count; i++)
             {
-                if (uniqueMatchedMaterials[i].Count != 1 || uniqueMatchedMaterials[i][0] == null)
+                if (uniqueMatchedMaterials[i].Count != 1 || uniqueMatchedMaterials[i][0].material == null)
                     continue;
                 var originalSlot = GetOriginalSlot((meshPath, matchedSlots[i][0].index));
                 AddAnimationPathChange((originalSlot.path, $"m_Materials.Array.data[{originalSlot.index}]", typeof(SkinnedMeshRenderer)),
@@ -4057,7 +4135,7 @@ public class d4rkAvatarOptimizer : MonoBehaviour
                 {
                     optimizedSlotSwapMaterials[originalSlot] = optimizedSwapMaterials = new Dictionary<Material, Material>();
                 }
-                optimizedSwapMaterials[uniqueMatchedMaterials[i][0]] = optimizedMaterials[i];
+                optimizedSwapMaterials[uniqueMatchedMaterials[i][0].material] = optimizedMaterials[i];
             }
 
             meshRenderer.sharedMaterials = optimizedMaterials;
@@ -4238,6 +4316,9 @@ public class d4rkAvatarOptimizer : MonoBehaviour
             var mergedMeshPaths = basicMergedMeshes.Select(list => list.Select(r => GetPathToRoot(r)).ToList()).ToList();
             basicMergedMeshesList.ForEach(r => oldPathToMergedPaths[GetPathToRoot(r)] = mergedMeshPaths);
             basicMergedMeshesList.ForEach(r => oldPathToMergedPath[GetPathToRoot(r)] = newPath);
+
+            // this caches the results for later use when optimizing the materials
+            basicMergedMeshesList.ForEach(r => GetFirstMaterialOnPath(GetPathToRoot(r)));
 
             foreach (SkinnedMeshRenderer skinnedMesh in basicMergedMeshesList)
             {